src/org/launchcode/techjobs_oo/Job.java
package org.launchcode.techjobs_oo;

import java.lang.reflect.Field;
import java.lang.reflect.Method;
import java.util.List;
import java.util.ListIterator;
import java.util.Objects;

import static java.util.Set.of;

public class Job {

    private int id;
    private static int nextId = 1;
    private int nullFields = 0;

    private String name;
    private Employer employer;
    private Location location;
    private PositionType positionType;
    private CoreCompetency coreCompetency;
    private Object o;

    // TODO: Add two constructors - one to initialize a unique ID and a second to initialize the
    //  other five fields. The second constructor should also call the first in order to initialize
    //  the 'id' field.

    public Job() {
        this.id = nextId;
        nextId++;
    }

    public Job(String name, Employer employer, Location location, PositionType positionType, CoreCompetency coreCompetency) {
        this();
        this.name = name;
        this.employer = employer;
        this.location = location;
        this.positionType = positionType;
        this.coreCompetency = coreCompetency;
    }

    private Object checkIfNull(Object o){
        if (o == null || o.toString() ==""){
            nullFields++;
            return "Data not available";
        } else
            return o;
    }

    // TODO: Add custom equals and hashCode methods. Consider two Job objects "equal" when their id fields
    //  match.

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Job job = (Job) o;
        return id == job.id;
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }

    @Override
    public String toString() {
        String str = "\n" +
                    "ID: "+ checkIfNull(id)+ "\n" +
                    "Name: "+checkIfNull(name)+ "\n" +
                    "Employer: "+checkIfNull(employer)+ "\n" +
                    "Location: "+checkIfNull(location)+ "\n" +
                    "Position Type: "+checkIfNull(positionType)+"\n" +
                    "Core Competency: "+checkIfNull(coreCompetency) +"\n" +
                    "";
        if(nullFields >= 5){
            return "OOPS! This job does not seem to exist.";
        }
        return str;
    }
    // TODO: Add getters for each field EXCEPT nextId. Add setters for each field EXCEPT nextID
    //  and id.

    public int getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Employer getEmployer() {
        return employer;
    }

    public void setEmployer(Employer employer) {
        this.employer = employer;
    }

    public Location getLocation() {
        return location;
    }

    public void setLocation(Location location) {
        this.location = location;
    }

    public PositionType getPositionType() {
        return positionType;
    }

    public void setPositionType(PositionType positionType) {
        this.positionType = positionType;
    }

    public CoreCompetency getCoreCompetency() {
        return coreCompetency;
    }

    public void setCoreCompetency(CoreCompetency coreCompetency) {
        this.coreCompetency = coreCompetency;
    }
}
